Implements dry-run mode; moves duplication of debug config

+++ chaos_agent/nodes.py
@@ -8,11 +8,11 @@ from chaos_agent.utils import configure_logging
 logger = configure_logging()
 
 
-def find_and_terminate_nodes():
+def find_and_terminate_nodes(dry_run):
     nodes = list_nodes()
     if nodes:
         node = select_random_node(nodes)
-        # delete_node(node)
+        # delete_node(node=node, dry_run=dry_run)
         logger.info(f"would delete {node} here")
 
 
@@ -32,10 +32,13 @@ def select_random_node(nodes):
     return nodes[node]
 
 
-def delete_node(node, grace=0):
-    core_api = client.CoreV1Api()
-    try:
-        logger.debug(f"Deleting node: {node}")
-        core_api.delete_node(node, grace_period_seconds=grace)
-    except ApiException as e:
-        logger.error(f"Failed to delete node, error message: {e}")
+def delete_node(node, dry_run=False, grace=0):
+    if dry_run:
+        logger.info(f"DRY-RUN: Node {node} would have been deleted")
+    else:
+        core_api = client.CoreV1Api()
+        try:
+            logger.debug(f"Deleting node: {node}")
+            core_api.delete_node(node, grace_period_seconds=grace)
+        except ApiException as e:
+            logger.error(f"Failed to delete node, error message: {e}")
